requirements.txt

python-telegram-bot==20.0a6
pandas
requests
openpyxl
APScheduler
nest_asyncio
